Mostly refinements: removed redundant index definitions; merged `gear` and `radios`. Details follow.
Changed `carriers`.`pattern`: removed NOT NULL and DEFAULT ''. This is because non-viable carriers will not have a pattern.

Removed redundant indices `id_UNIQUE`. The name of the primary key is `PRIMARY` and is defined by the statement PRIMARY KEY (`id`). It is not necessary to specify the ID as a unique key after making that statement.

Removed NOT NULL from `pairs`.`team`, on the suspicion it is possible for ops who are not part of a team to be in a pair. (E.g. MODs.)

Changed all TINYINT(1) column definitions to use SET instead, with one value. `permission_groups`.`canLogIn` became `permission_groups`.`permissions`, with 'canLogIn' as a SET element. `gear`.`broken` became `gear`.`flags`, with 'broken' as an element.

Dropped `radios` and `radio_transactions` table, and added `channel` and its relevant foreign key constraint to `gear`.

Changed constraint `gear_owner` to ON DELETE RESTRICT. I don't think we should be discarding information tying equipment to operatives, and that's exactly what ON DELETE SET NULL was going to do if an op were deleted. (Discuss use cases for deleting ops. It's possible to add a "hidden" flag for ops so that my procedures don't capture them, if desired.) Wisely, the `gear_transactions` table already used ON DELETE RESTRICT.

In relationship tables `offenders`, `incident_ops`, `victims`, and `witnesses`, removed indexes which were left prefixes of the primary key. These were redundant.

Removed NOT NULL from `log_entries`.`channel`. The log can be used for events which were not related to Dispatch using radio traffic.

Disabled column `log_entries`.`onShift` because I think the current state of the spec does not make sense. Discussion needed.